backend/authentication/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
blis==0.7.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.5
cymem==2.0.7
Django==4.2.1
django-cors-headers==4.1.0
django-rest-registration==0.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
en-core-web-sm==3.5.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
murmurhash==1.0.9
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pathy==0.10.2
Pillow==9.5.0
preshed==3.0.8
psycopg2==2.9.6
pydantic==1.10.10
PyJWT==2.7.0
pytesseract==0.3.10
python-decouple==3.8
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
smart-open==6.3.0
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.4
srsly==2.4.6
thinc==8.1.10
tqdm==4.65.0
typer==0.9.0
typing-extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
wasabi==1.1.2
